add support for exports and exports_sources as methods

+++ conans/client/cmd/export.py
@@ -10,6 +10,7 @@ from conans.client.file_copier import FileCopier
 from conans.client.output import Color, ScopedOutput
 from conans.client.remover import DiskRemover
 from conans.errors import ConanException, ConanV2Exception
+from conans.model.build_info import DepsCppInfo
 from conans.model.manifest import FileTreeManifest
 from conans.model.ref import ConanFileReference
 from conans.model.scm import SCM, get_scm_data
@@ -447,41 +448,51 @@ def _export_scm(scm_data, origin_folder, scm_sources_folder, output):
 
 
 def export_source(conanfile, origin_folder, destination_source_folder):
-    if isinstance(conanfile.exports_sources, str):
-        conanfile.exports_sources = (conanfile.exports_sources, )
+    if callable(conanfile.exports_sources):
+        conanfile.deps_cpp_info = DepsCppInfo()
+        conanfile.copy = FileCopier([origin_folder], destination_source_folder)
+        conanfile.export_sources()
+    else:
+        if isinstance(conanfile.exports_sources, str):
+            conanfile.exports_sources = (conanfile.exports_sources, )
 
-    included_sources, excluded_sources = _classify_patterns(conanfile.exports_sources)
-    copier = FileCopier([origin_folder], destination_source_folder)
-    for pattern in included_sources:
-        copier(pattern, links=True, excludes=excluded_sources)
-    output = conanfile.output
-    package_output = ScopedOutput("%s exports_sources" % output.scope, output)
-    copier.report(package_output)
+        included_sources, excluded_sources = _classify_patterns(conanfile.exports_sources)
+        copier = FileCopier([origin_folder], destination_source_folder)
+        for pattern in included_sources:
+            copier(pattern, links=True, excludes=excluded_sources)
+        output = conanfile.output
+        package_output = ScopedOutput("%s exports_sources" % output.scope, output)
+        copier.report(package_output)
 
 
 def export_recipe(conanfile, origin_folder, destination_folder):
-    if isinstance(conanfile.exports, str):
-        conanfile.exports = (conanfile.exports, )
+    if callable(conanfile.exports):
+        conanfile.deps_cpp_info = DepsCppInfo()
+        conanfile.copy = FileCopier([origin_folder], destination_folder)
+        conanfile.exports()
+    else:
+        if isinstance(conanfile.exports, str):
+            conanfile.exports = (conanfile.exports, )
 
-    output = conanfile.output
-    package_output = ScopedOutput("%s exports" % output.scope, output)
+        output = conanfile.output
+        package_output = ScopedOutput("%s exports" % output.scope, output)
 
-    if os.path.exists(os.path.join(origin_folder, DATA_YML)):
-        package_output.info("File '{}' found. Exporting it...".format(DATA_YML))
-        tmp = [DATA_YML]
-        if conanfile.exports:
-            tmp.extend(conanfile.exports)  # conanfile.exports could be a tuple (immutable)
-        conanfile.exports = tmp
+        if os.path.exists(os.path.join(origin_folder, DATA_YML)):
+            package_output.info("File '{}' found. Exporting it...".format(DATA_YML))
+            tmp = [DATA_YML]
+            if conanfile.exports:
+                tmp.extend(conanfile.exports)  # conanfile.exports could be a tuple (immutable)
+            conanfile.exports = tmp
 
-    included_exports, excluded_exports = _classify_patterns(conanfile.exports)
+        included_exports, excluded_exports = _classify_patterns(conanfile.exports)
 
-    try:
-        os.unlink(os.path.join(origin_folder, CONANFILE + 'c'))
-    except OSError:
-        pass
+        try:
+            os.unlink(os.path.join(origin_folder, CONANFILE + 'c'))
+        except OSError:
+            pass
 
-    copier = FileCopier([origin_folder], destination_folder)
-    for pattern in included_exports:
-        copier(pattern, links=True, excludes=excluded_exports)
+        copier = FileCopier([origin_folder], destination_folder)
+        for pattern in included_exports:
+            copier(pattern, links=True, excludes=excluded_exports)
 
-    copier.report(package_output)
+        copier.report(package_output)
